SystemUI: Fix display cutout handling in expanding volume panel
Devices with display cutouts in the status bar area usually have a
taller status bar to accomodate them. The custom volume panel dialog
insets don't account for that when using the location of the dialog,
resulting in incorrect insets that block touch at the top edge of the
volume panel.

Custom insets are unnecessary anyway, especially when they just return a
rectangle encompassing the entire visible region of the volume panel, so
remove the custom inset code to fix the display cutout handling issue.

Collapsed volume controls aren't touchable regardless of insets because
they're inside the collapsed container.

Change-Id: I6b8f2f796a99c7d28c0660988e208a5d831379b6
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: LordShenron <shen.priyanshu@gmail.com>
Signed-off-by: Aryan Sinha <sinha.aryan03@gmail.com>